// src/app/modules/barbara-damage-viewer/config/inspekt-damages.enum.ts
export enum InspektDamages {
    SCRATCH_OR_SPOT = 'scratch_or_spot',
    SHALLOW_DENT = 'shallow_dent',
    DEEP_DENT = 'deep_dent',
    DESIGN_DENT = 'design_dent',
    TEAR = 'tear',
    DISLOCATION = 'dislocation',
    SHATTER = 'shatter',
    BROKEN = 'broken',
    DISLOCATION_2 = 'dislocation_2',
    STITCH_OR_SCREW = 'stitch_or_screw',
    RUST = 'rust',
    COLLISION_DAMAGE = 'collision_damage',
    MISSING_PART = 'missing_part',
    CRACK = 'crack',
    HOLE = 'hole',
    CHIP_OFF = 'chip_off',
    SPIDER_CORE = 'spider_core',
    SPREAD = 'spread',
    AI_UNCERTAIN = 'AI uncertain',
    SPOT_OR_DIRT = 'spot_or_dirt',
    //
    SPOT = 'spot',
    SHALLOW_DENT_MINOR = 'shallow_dent_minor',
    SHALLOW_DENT_MAJOR = 'shallow_dent_major',
    SCRATCH = 'scratch',
    SCRATCH_MINOR = 'scratch_minor',
    SCRATCH_MAJOR = 'scratch_major',
    DESIGN_DENT_MINOR = 'design_dent_minor',
    DESIGN_DENT_MAJOR = 'design_dent_major',
    RUST_EXTERNAL = 'rust_external'
}